refactor

+++ src/query-performance-monitoring/datamodels/query_datamodels.go
@@ -42,9 +42,9 @@ type IndividualQuerySearch struct {
 }
 
 type QueryCpuMetrics struct {
-	queryId      *int64  `db:"queryid" metric_name:"querycpu.queryid" source_type:"gauge"`
-	AvgCpuTime   *int64  `db:"avg_cpu_time" metric_name:"querycpu.avg_cpu_time" source_type:"gauge"`
-	DatabaseName *string `db:"datname" metric_name:"querycpu.database" source_type:"attribute"`
+	queryId      *int64   `db:"queryid" metric_name:"querycpu.queryid" source_type:"gauge"`
+	AvgCpuTime   *float64 `db:"avg_cpu_time" metric_name:"querycpu.avg_cpu_time" source_type:"gauge"`
+	DatabaseName *string  `db:"datname" metric_name:"querycpu.database" source_type:"attribute"`
 }
 
 type QueryExecutionPlanMetrics struct {
